edge/sensorManager/pyproject.toml
[tool.poetry]
name = "w4b-sensor-manager"
version = "0.1.0"
description = "Sensor management system for the w4b beehive monitoring platform"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "sensorManager"}]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0"
asyncpg = "^0.27.0"
prometheus-client = "^0.16.0"
jsonschema = "^4.17.3"
python-dotenv = "^0.21.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
black = "^23.3.0"
mypy = "^1.2.0"
flake8 = "^6.0.0"
bandit = "^1.7.5"
pytest-cov = "^4.1.0"
pylint = "^2.17.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.poetry.scripts]
sensor-manager = "sensorManager.main:main_cli"
